% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-container.R
\name{create_shinylive_container}
\alias{create_shinylive_container}
\title{Create a container for shinylive application}
\usage{
create_shinylive_container(
  app_path,
  id,
  mode = c("viewer", "editor-terminal-viewer"),
  height = NULL
)
}
\arguments{
\item{id}{Unique container id.}

\item{mode}{shinylive app configuration. Display either the app, the app +
code editor + terminal.}

\item{height}{Container height. Expect a valid CSS unit like \verb{800px}.}

\item{app_json_path}{app_path App location. We call shinylive::export
to create a JSON file which is then forwarded to shinylive. This only works
if your app is within a directory and not created from a R function.}
}
\value{
A tag list.
}
\description{
Allows to embed a shinylive standalone app into a pkgdown website.
Importantly, you need to serve the website to preview the applications.
This won't work with the classic pkgdown local browser. The best way to do so
is either to use \code{httpuv::runStaticServer("<SITE_PATH>")} or use the
\verb{go live} VSCode extension.
}
